requirements.txt
alembic==1.8.1
aniso8601==9.0.1
astroid==2.3.1
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
confection==0.0.3
contourpy==1.0.5
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
dnspython==2.3.0
executing==1.1.1
filelock==3.8.0
Flask==2.1.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
fonttools==4.38.0
googletrans==4.0.0rc1
greenlet==1.1.3.post0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.10.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.10.1
hyperframe==5.2.0
idna==2.10
importlib-metadata==5.0.0
importlib-resources==5.10.1
ipython==8.5.0
isort==4.3.21
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy-object-proxy==1.4.2
llvmlite==0.39.1
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mccabe==0.6.1
murmurhash==1.0.9
numba==0.56.3
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathy==0.6.2
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
preshed==3.0.8
prompt-toolkit==3.0.31
pure-eval==0.2.2
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.4.2
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.1
pytz==2022.5
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
rfc3986==1.5.0
scikit-learn==1.1.3
scipy==1.9.3
sentencepiece==0.1.97
shap==0.41.0
six==1.12.0
slicer==0.0.7
smart-open==5.2.1
sniffio==1.3.0
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.42
srsly==2.4.5
stack-data==0.5.1
thinc==8.1.5
threadpoolctl==3.1.0
tokenizers==0.13.1
torch==1.12.1
tqdm==4.64.1
traitlets==5.5.0
transformers==4.23.1
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.11.2
zipp==3.10.0

gunicorn